requirements.txt
tensorflow>=2.10.0
opencv-python>=4.7.0
mediapipe>=0.10.0
numpy>=1.23.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pygame>=2.5.0
pandas>=2.0.0
Pillow>=9.5.0
torch>=2.11.0
transformers>=4.40.0
huggingface_hub>=0.23.0
